iuser-main/requirements.txt
behave==1.2.6
google-api-python-client==2.9.0
google-cloud-bigquery==2.13.1
google-cloud-core==1.6.0
google-cloud-datastore==1.15.3
google-cloud-logging==2.3.1
google-cloud-ndb==1.8.0
google-cloud-pubsub==2.4.1
google-cloud-storage==1.37.1
matplotlib==3.4.1
mysql-connector-python==8.0.25
pandas-gbq==0.15.0
pandas==1.2.3
psutil==5.8.0
pytest-cov==2.12.1
pytest==6.2.4
python-dotenv==0.17.1
requests==2.25.1
WebTest==2.0.35
WSGIProxy2==0.4.6